+++ src/components/Preloader/Preloader.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { useNews } from "../../contexts/NewsContext";
+import { useNews } from "../../contexts/GlobalContext";
 import "./Preloader.css";
 
 function Preloader() {
